reuse prerequisiteMaterial component
